requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
sse-starlette==1.6.5
python-multipart==0.0.6

# Async HTTP Client
aiohttp==3.9.0
httpx==0.25.2

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration & Environment
python-dotenv==1.0.0

# Retry Logic & Resilience
tenacity==8.2.3

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Monitoring & Metrics
prometheus-client==0.19.0

# Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# Additional utilities
asyncio-mqtt==0.16.1
pytz==2023.3
